Kernel Current I0 factor to float and randomness removed
StdDev set to zero to remove randomness (stupid way because the random
number generation is still eating CPU time)
I0 factor is now being multiplied during NeuronSimulate instead of
CurrentUpdate so that it can be a float. In storage, Iins are still
integers but Iext and Itot will be in Ampere (float).