Reintroducing SpikeQueue usage
On spike arrival, the SpikeQueue (a circular queue of length EPSP
length) is updated with synapses that are going to be ON at the time
pointed by the CurrentQueueIndex.
1. CurrentAttenuate makes current zero (each kernel can be thought of as
a dirac delta pulse and EPSP length of such pulses are created)
2. NeuronSimulate adds EPSP length spikes to the buffer
3. SpikeRecord was changed to allow updates described above.

Problem: Thsi treats all spikes equivalently. In the paper, consecutive
spikes from the same neuron with EPSP time difference just expand the
EPSP window. Here, they will add more current. So, at each index of the
SpikeQueue, one synapse should be present only once.